Pipeline cleaned
