--- genres.csv
@@ -1,23 +1,0 @@
-Metal
-Experimental
-Ambient
-Pop
-Rock
-Alternative
-Rap
-Hip Hop
-Disco
-Dance
-Electronic
-Techno
-Jazz
-Soul
-Punk
-DNB
-Funk
-Dub
-Folk
-Indie
-House
-R&B
-Country
